refactor(common): move common components to dedicated directory with aliases
